python/pyproject.toml
[project]
name = "kadoa-sdk"
version = "0.2.1"
description = "Kadoa official python SDK"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "websocket-client>=1.8.0",
]

[tool.setuptools]
packages = ["kadoa_sdk"]

[dependency-groups]
dev = [
    "bump2version>=1.0.1",
    "pysemver>=0.5.0",
    "pytest>=8.3.4",
    "semver>=3.0.2",
    "twine>=6.0.1",
]

[tool.pytest.ini_options]
pythonpath = [
  "."
]
